better table borders
git-svn-id: svn://libnui.net/nui/trunk/nui3@5273 9db30882-7939-4ce0-a36a-45f8a15c01b5

+++ src/Text/HTML/nuiHTMLTable.h
@@ -84,7 +84,7 @@ public:
     nglString mAbbr;
     bool mHeader;
     Frame mFrame;
-    float mBorder;
+    int32 mBorder;
     float mSpacing;
     float mPadding;
   };
